[svn-r12706] Description:     Clean up some of the warnings on 64-bit Linux...
Tested on:
    Linux/32 2.6 (chicago)
    Linux/64 2.6 (chicago2)
    Too minor to require h5committest

+++ src/H5C.c
@@ -3458,7 +3458,7 @@ H5C_flush_to_min_clean(H5F_t * f,
                                      primary_dxpl_id,
                                      secondary_dxpl_id,
                                      cache_ptr,
-                                     0,
+                                     (size_t)0,
                                      write_permitted,
                                      &first_flush);
